+++ cli/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+
+kubernetes.ingress.class=nginx2
