+++ cli/src/test/java/fr/insee/innovation/marathontokubernetes/cli/MarathonToKubernetesJavaCliApplicationTests.java
@@ -1,0 +1,13 @@
+package fr.insee.innovation.marathontokubernetes.cli;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MarathonToKubernetesJavaCliApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
