+++ core/src/main/java/fr/insee/innovation/marathontokubernetes/core/MarathonToKubernetesApplication.java
