+++ core/src/main/java/fr/insee/innovation/marathontokubernetes/core/configuration/CustomObjectMapper.java
