+++ core/src/main/java/fr/insee/innovation/marathontokubernetes/core/configuration/PropertiesLoader.java
@@ -1,0 +1,9 @@
+package fr.insee.innovation.marathontokubernetes.core.configuration;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+
+@Configuration
+@PropertySource("classpath:common.properties")
+public class PropertiesLoader {
+}
