+++ core/src/main/resources/common.properties
