+++ core/src/test/java/fr/insee/innovation/marathontokubernetes/core/MarathonToKubernetesApplicationTests.java
