+++ core/src/test/resources/marathon/drawio.json
