+++ core/src/test/resources/marathon/privileged.json
