--- marathon-to-kubernetes-core/pom.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" 
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<artifactId>marathon-to-kubernetes-core</artifactId>
-
-	<parent>
-		<groupId>fr.insee.innovation</groupId>
-		<artifactId>marathon-to-kubernetes-parent</artifactId>
-		<version>0.0.1-SNAPSHOT</version>
-		<relativePath>../pom.xml</relativePath>
-	</parent>
-
-	<name>marathon-to-kubernetes-core</name>
-	<description>Convert marathon.json to kubernetes yml</description>
-
-	<properties>
-		<java.version>11</java.version>
-	</properties>
-
-	<dependencies>
-
-		<dependency>
-			<groupId>io.fabric8</groupId>
-			<artifactId>kubernetes-client</artifactId>
-			<version>4.6.4</version>
-		</dependency>
-
-		<!-- https://mvnrepository.com/artifact/com.mesosphere/marathon-client -->
-		<dependency>
-			<groupId>com.mesosphere</groupId>
-			<artifactId>marathon-client</artifactId>
-			<version>0.6.2</version>
-		</dependency>
-
-	</dependencies>
-
-	<build>
-
-	</build>
-
-</project>
