--- marathon-to-kubernetes-java-cli/pom.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" 
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<artifactId>marathon-to-kubernetes-java-cli</artifactId>
-
-	<parent>
-		<groupId>fr.insee.innovation</groupId>
-		<artifactId>marathon-to-kubernetes-parent</artifactId>
-		<version>0.0.1-SNAPSHOT</version>
-		<relativePath>../pom.xml</relativePath>
-	</parent>
-
-
-	<name>marathon-to-kubernetes-java-cli</name>
-	<description>Convert marathon.json to kubernetes yml</description>
-
-	<properties>
-		<java.version>11</java.version>
-	</properties>
-
-	<dependencies>
-		<dependency>
-			<groupId>fr.insee.innovation</groupId>
-			<artifactId>marathon-to-kubernetes-core</artifactId>
-			<version>0.0.1-SNAPSHOT</version>
-		</dependency>
-		
-		<dependency>
-			<groupId>info.picocli</groupId>
-			<artifactId>picocli-spring-boot-starter</artifactId>
-			<version>4.1.2</version>
-		</dependency>
-
-	</dependencies>
-
-	<build>
-
-	</build>
-
-</project>
