--- marathon-to-kubernetes-java-cli/src/test/java/fr/insee/innovation/marathontokubernetesjavacli/MarathonToKubernetesJavaCliApplicationTests.java
@@ -1,13 +1,0 @@
-package fr.insee.innovation.marathontokubernetesjavacli;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MarathonToKubernetesJavaCliApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
